+++ crates/notan_winit/src/lib.rs
@@ -5,6 +5,7 @@ mod mouse;
 mod touch;
 mod window;
 
+mod gl_manager;
 pub mod prelude;
 
 pub use backend::*;
